added error msg when ether is not enough

+++ controllers/documentController.js
@@ -12,7 +12,7 @@ const sendEmail = require('../utils/sendEmail');
 
 //Dependencies needed to deploy and interact with the smart contract
 const deployContract = require('../contracts/pseudoDeploy');
-const {newCertificate, getAllCertificates} = require('../contracts/interact');
+const {newCertificate, getAllCertificates, getWalletBalance} = require('../contracts/interact');
 const getTransactionData = require('../contracts/getTransactionData');
 
 //======================
@@ -58,6 +58,7 @@ router.post('/certificates', requireAuth, async(req, res) => {
         res.status(200).json({data: "Success", result: result});
     } catch (err) {
         console.log(err);
+        res.status(404).json({dataError: "Error at backend"});
     }
     
 
@@ -69,19 +70,27 @@ router.post('/deploy', requireAuth, async(req,res) => {
     if (req.profile.id === req.body.userId) {
         try {
             const user = await User.findOne({_id: req.body.userId});
-            const {address} = web3.eth.accounts.privateKeyToAccount(req.body.privateKey);                    
-            const contractAddress = await deployContract(testnetObj[req.body.testnet], address, req.body.privateKey);
+            const {address} = web3.eth.accounts.privateKeyToAccount(req.body.privateKey);     
+            const etherBalance = await getWalletBalance(testnetObj[req.body.testnet], address, req.body.privateKey);
             
-            const ethNet = await EthereumNet.create({
-                userId: req.body.userId,
-                nameOfNet: req.body.testnet,
-                address: contractAddress
-            });
+            if (parseFloat(etherBalance) >= 0.01) {
+                const contractAddress = await deployContract(testnetObj[req.body.testnet], address, req.body.privateKey);
             
-            res.status(200).json({data: "Success"});
+                const ethNet = await EthereumNet.create({
+                    userId: req.body.userId,
+                    nameOfNet: req.body.testnet,
+                    address: contractAddress
+                });
+                
+                res.status(200).json({data: "Success"});
+            } else {
+                console.log("Not enough Ether", etherBalance);
+                res.status(404).json({dataError: `You do not have at least 0.01 Ether to deploy Contract. You only have ${etherBalance} Ether. Please get more Ether`});
+            }
+            
         } catch (err) {
             console.log(err);
-            res.status(404).json({dataError: err})
+            res.status(404).json({dataError: "Error at backend"})
         }
           
     }
@@ -116,16 +125,23 @@ router.post('/new', requireAuth, async(req, res) => {
             };
 
             const {address} = web3.eth.accounts.privateKeyToAccount(req.body.privateKey);
-            const transactionReceipt = await newCertificate(testnetObj[req.body.testnet], address, req.body.privateKey, deployedContract.address, certificateParams);
-            const rootPath = `${req.protocol}://${req.hostname}:${process.env.REACT_PORT}/documents/${req.body.testnet}`; 
-            await sendEmail(req.body.studentEmail, req.body.studentName, transactionReceipt.transactionHash, file, rootPath);
-
-            fs.unlinkSync(`./uploads/${file.filename}`);            
+            const etherBalance = await getWalletBalance(testnetObj[req.body.testnet], address, req.body.privateKey);
             
-            res.status(200).json({data: "Success"});
+            if (parseFloat(etherBalance) >= 0.01) {
+                const transactionReceipt = await newCertificate(testnetObj[req.body.testnet], address, req.body.privateKey, deployedContract.address, certificateParams);
+                const rootPath = `${req.protocol}://${req.hostname}:${process.env.REACT_PORT}/documents/${req.body.testnet}`; 
+                await sendEmail(req.body.studentEmail, req.body.studentName, transactionReceipt.transactionHash, file, rootPath);
+
+                fs.unlinkSync(`./uploads/${file.filename}`);            
+                
+                res.status(200).json({data: "Success"});
+            } else {
+                res.status(404).json({dataError: `You do not have at least 0.01 Ether. You only have ${etherBalance} Ether. Please get more Ether`});
+            }
         }  
     } catch(err) {
         console.log(err);
+        res.status(404).json({dataError: "Error at backend"});
     }
 });
 
@@ -148,6 +164,7 @@ router.post('/verifyIpfs', async (req,res) => {
         }
     } catch (err) {
         console.log(err);
+        res.status(404).json({dataError: "Error at backend"});
     }
    
 
@@ -178,5 +195,17 @@ router.get('/verify/:testnet/:txnHash', async (req, res) => {
     }
 });
 
+//const getWalletBalance = async(testnetArg, addressArg, privateKeyArg) => {   
+// router.get('/walletBalance', async(req,res) => {
+//     try {
+//         const {address} = web3.eth.accounts.privateKeyToAccount("48d29642ec637702fe6e5017dacc71cc75f4b9df5df8f5114901f89fc46a862f");                    
+//         const etherBalance = await getWalletBalance(testnetObj["localhost"], address, "48d29642ec637702fe6e5017dacc71cc75f4b9df5df8f5114901f89fc46a862f");
+//         res.status(200).json({data: etherBalance});
+//     } catch (err) {
+//         console.log(err);
+//     }
+    
+
+// });
 
 module.exports = router;
